+++ engine/standalone/api/src/main/scala/pl/touk/nussknacker/engine/standalone/api/StandaloneContextLifecycle.scala
@@ -1,0 +1,12 @@
+package pl.touk.nussknacker.engine.standalone.api
+
+import pl.touk.nussknacker.engine.api.JobData
+
+trait StandaloneContextLifecycle {
+
+  def open(jobData: JobData, context: StandaloneContext): Unit
+
+  def close(): Unit
+
+
+}
