+++ engine/standalone/app/src/main/scala/pl/touk/nussknacker/engine/standalone/http/StandaloneHttpApp.scala
@@ -8,11 +8,12 @@ import com.typesafe.config.{Config, ConfigFactory}
 import com.typesafe.scalalogging.LazyLogging
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
 import io.dropwizard.metrics5.MetricRegistry
+import pl.touk.nussknacker.engine.standalone.api.StandaloneContextPreparer
 import pl.touk.nussknacker.engine.standalone.deployment.DeploymentService
-import pl.touk.nussknacker.engine.standalone.utils.StandaloneContextPreparer
-import pl.touk.nussknacker.engine.standalone.utils.logging.StandaloneRequestResponseLogger
-import pl.touk.nussknacker.engine.standalone.utils.metrics.dropwizard.influxdb.StandaloneInfluxDbReporter
-import pl.touk.nussknacker.engine.standalone.utils.metrics.dropwizard.{DropwizardMetricsProvider, StandaloneMetricsReporter}
+import pl.touk.nussknacker.engine.standalone.http.logging.StandaloneRequestResponseLogger
+import pl.touk.nussknacker.engine.standalone.http.metrics.dropwizard.influxdb.StandaloneInfluxDbReporter
+import pl.touk.nussknacker.engine.standalone.http.metrics.dropwizard.StandaloneMetricsReporter
+import pl.touk.nussknacker.engine.standalone.metrics.dropwizard.DropwizardMetricsProvider
 import pl.touk.nussknacker.engine.util.loader.ScalaServiceLoader
 
 import scala.util.Try
