+++ engine/standalone/app/src/main/scala/pl/touk/nussknacker/engine/standalone/http/metrics/dropwizard/influxdb/InfluxDbHttpReporter.scala
@@ -1,4 +1,4 @@
-package pl.touk.nussknacker.engine.standalone.utils.metrics.dropwizard.influxdb
+package pl.touk.nussknacker.engine.standalone.http.metrics.dropwizard.influxdb
 
 import com.typesafe.scalalogging.LazyLogging
 import io.dropwizard.metrics5.influxdb._
