+++ engine/standalone/app/src/test/scala/pl/touk/nussknacker/engine/standalone/http/TestConfigCreator.scala
@@ -7,7 +7,7 @@ import io.circe.generic.JsonCodec
 import pl.touk.nussknacker.engine.api.process._
 import pl.touk.nussknacker.engine.api.{JobData, MethodToInvoke, Service}
 import pl.touk.nussknacker.engine.standalone.api.types.GenericResultType
-import pl.touk.nussknacker.engine.standalone.api.{ResponseEncoder, StandaloneGetSource, StandaloneSourceFactory}
+import pl.touk.nussknacker.engine.standalone.api.{ResponseEncoder, StandaloneContext, StandaloneContextLifecycle, StandaloneGetSource, StandaloneSinkFactory, StandaloneSourceFactory}
 import pl.touk.nussknacker.engine.standalone.utils._
 import pl.touk.nussknacker.engine.util.json.BestEffortJsonEncoder
 import pl.touk.nussknacker.engine.util.process.EmptyProcessConfigCreator
