+++ engine/standalone/engine/src/main/scala/pl/touk/nussknacker/engine/standalone/StandaloneProcessInterpreter.scala
@@ -23,9 +23,8 @@ import pl.touk.nussknacker.engine.resultcollector.ResultCollector
 import pl.touk.nussknacker.engine.split.{NodesCollector, ProcessSplitter}
 import pl.touk.nussknacker.engine.splittedgraph.splittednode.SplittedNode
 import pl.touk.nussknacker.engine.standalone.api.types._
-import pl.touk.nussknacker.engine.standalone.api.{JoinStandaloneCustomTransformer, StandaloneCustomTransformer, StandaloneSource, types}
+import pl.touk.nussknacker.engine.standalone.api._
 import pl.touk.nussknacker.engine.standalone.metrics.InvocationMetrics
-import pl.touk.nussknacker.engine.standalone.utils.{StandaloneContext, StandaloneContextLifecycle, StandaloneContextPreparer, StandaloneSinkWithParameters}
 import pl.touk.nussknacker.engine.{ModelData, compiledgraph}
 
 import scala.concurrent.{ExecutionContext, Future}
