--- engine/standalone/util/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/StandaloneContextLifecycle.scala
@@ -1,12 +1,0 @@
-package pl.touk.nussknacker.engine.standalone.utils
-
-import pl.touk.nussknacker.engine.api.JobData
-
-trait StandaloneContextLifecycle {
-
-  def open(jobData: JobData, context: StandaloneContext) : Unit
-
-  def close() : Unit
-
-
-}
