+++ engine/standalone/util/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/metrics/WithEspTimers.scala
@@ -1,7 +1,7 @@
 package pl.touk.nussknacker.engine.standalone.utils.metrics
 
 import cats.data.NonEmptyList
-import pl.touk.nussknacker.engine.standalone.utils.StandaloneContext
+import pl.touk.nussknacker.engine.standalone.api.StandaloneContext
 import pl.touk.nussknacker.engine.util.service.EspTimer
 
 trait WithEspTimers {
@@ -13,7 +13,3 @@ trait WithEspTimers {
   def espTimer(tags: Map[String, String], name: NonEmptyList[String]): EspTimer = context.espTimer(instantTimerWindowInSeconds, tags, name)
 
 }
-
-
-
-
