+++ engine/standalone/util/src/main/scala/pl/touk/nussknacker/engine/standalone/utils/service/TimeMeasuringService.scala
@@ -1,8 +1,8 @@
 package pl.touk.nussknacker.engine.standalone.utils.service
 
 import pl.touk.nussknacker.engine.api.{JobData, Service}
+import pl.touk.nussknacker.engine.standalone.api.{StandaloneContext, StandaloneContextLifecycle}
 import pl.touk.nussknacker.engine.standalone.utils.metrics.WithEspTimers
-import pl.touk.nussknacker.engine.standalone.utils.{StandaloneContext, StandaloneContextLifecycle}
 import pl.touk.nussknacker.engine.util.service.GenericTimeMeasuringService
 
 trait TimeMeasuringService extends GenericTimeMeasuringService with StandaloneContextLifecycle with WithEspTimers { self: Service =>
